# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Lokacara API",
        "description": "API documentation for the Lokacara event management platform. Use the Authorize button at the top to set your Sanctum Bearer token for authenticated endpoints.",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/organizer/events": {
            "get": {
                "tags": [
                    "Organizer"
                ],
                "summary": "List my events (paginated)",
                "operationId": "d02476d69bef2ed2d2da2b02008e1711",
                "responses": {
                    "200": {
                        "description": "Paginated list of organizer events",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Event"
                                            }
                                        },
                                        "current_page": {
                                            "type": "integer"
                                        },
                                        "last_page": {
                                            "type": "integer"
                                        },
                                        "per_page": {
                                            "type": "integer"
                                        },
                                        "total": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Organizer"
                ],
                "summary": "Create a new event",
                "operationId": "69f033808abaa83c537d68855555e5fd",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "title",
                                    "description",
                                    "type",
                                    "start_datetime",
                                    "end_datetime",
                                    "poster"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Summer Music Festival",
                                        "maxLength": 255
                                    },
                                    "category_id": {
                                        "type": "integer",
                                        "example": 1,
                                        "nullable": true
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "A great music event with live performances"
                                    },
                                    "type": {
                                        "type": "string",
                                        "example": "offline",
                                        "enum": [
                                            "online",
                                            "offline"
                                        ]
                                    },
                                    "location_name": {
                                        "type": "string",
                                        "example": "Central Park"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "123 Main St, City"
                                    },
                                    "latitude": {
                                        "type": "number",
                                        "format": "float",
                                        "example": -6.2088
                                    },
                                    "longitude": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 106.8456
                                    },
                                    "platform_name": {
                                        "type": "string",
                                        "example": "Zoom"
                                    },
                                    "link": {
                                        "type": "string",
                                        "format": "uri",
                                        "example": "https://zoom.us/j/123456"
                                    },
                                    "start_datetime": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2026-07-15T09:00:00"
                                    },
                                    "end_datetime": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2026-07-15T17:00:00"
                                    },
                                    "capacity": {
                                        "type": "integer",
                                        "example": 100
                                    },
                                    "poster": {
                                        "description": "Event poster image (max 5MB, jpeg/png/webp)",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Event created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Event created successfully"
                                        },
                                        "event": {
                                            "$ref": "#/components/schemas/Event"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/organizer/events/{event}": {
            "post": {
                "tags": [
                    "Organizer"
                ],
                "summary": "Update an existing event",
                "operationId": "02bc3878e2df361b4c6f562164599bf8",
                "parameters": [
                    {
                        "name": "event",
                        "in": "path",
                        "description": "Event ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Updated Festival Name",
                                        "maxLength": 255
                                    },
                                    "category_id": {
                                        "type": "integer",
                                        "example": 1,
                                        "nullable": true
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Updated description"
                                    },
                                    "type": {
                                        "type": "string",
                                        "example": "online",
                                        "enum": [
                                            "online",
                                            "offline"
                                        ]
                                    },
                                    "platform_name": {
                                        "type": "string",
                                        "example": "Google Meet"
                                    },
                                    "link": {
                                        "type": "string",
                                        "format": "uri",
                                        "example": "https://meet.google.com/abc-defg-hij"
                                    },
                                    "start_datetime": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2026-08-01T10:00:00"
                                    },
                                    "end_datetime": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2026-08-01T18:00:00"
                                    },
                                    "capacity": {
                                        "type": "integer",
                                        "example": 150
                                    },
                                    "poster": {
                                        "description": "New poster image (optional)",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Event updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Event updated successfully"
                                        },
                                        "event": {
                                            "$ref": "#/components/schemas/Event"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden (not the owner)"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Organizer"
                ],
                "summary": "Delete an event (owner or admin)",
                "operationId": "fdec5865c15b3654427467aba928496d",
                "parameters": [
                    {
                        "name": "event",
                        "in": "path",
                        "description": "Event ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Event deleted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Event deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/organizer/events/{event}/attendees": {
            "get": {
                "tags": [
                    "Organizer"
                ],
                "summary": "Get attendees list for an event",
                "operationId": "88cd63b30d396bb72cf1d6bef1e18f7d",
                "parameters": [
                    {
                        "name": "event",
                        "in": "path",
                        "description": "Event ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Attendees list",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "event": {
                                            "$ref": "#/components/schemas/Event"
                                        },
                                        "attendees": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden (not the owner)"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Category": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Music"
                    },
                    "slug": {
                        "type": "string",
                        "example": "music"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
            "Event": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "category_id": {
                        "type": "integer",
                        "example": 2,
                        "nullable": true
                    },
                    "type": {
                        "type": "string",
                        "example": "offline",
                        "enum": [
                            "online",
                            "offline"
                        ]
                    },
                    "poster": {
                        "type": "string",
                        "example": "posters/abc123.jpg",
                        "nullable": true
                    },
                    "poster_url": {
                        "type": "string",
                        "example": "http://localhost:8000/api/posters/abc123.jpg",
                        "nullable": true
                    },
                    "title": {
                        "type": "string",
                        "example": "Summer Music Festival"
                    },
                    "description": {
                        "type": "string",
                        "example": "A great music event"
                    },
                    "location_name": {
                        "type": "string",
                        "example": "Central Park",
                        "nullable": true
                    },
                    "address": {
                        "type": "string",
                        "example": "123 Main St, City",
                        "nullable": true
                    },
                    "latitude": {
                        "type": "number",
                        "format": "float",
                        "example": -6.2088,
                        "nullable": true
                    },
                    "longitude": {
                        "type": "number",
                        "format": "float",
                        "example": 106.8456,
                        "nullable": true
                    },
                    "platform_name": {
                        "type": "string",
                        "example": "Zoom",
                        "nullable": true
                    },
                    "link": {
                        "type": "string",
                        "example": "https://zoom.us/j/123456",
                        "nullable": true
                    },
                    "start_datetime": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-07-15T09:00:00Z"
                    },
                    "end_datetime": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-07-15T17:00:00Z"
                    },
                    "capacity": {
                        "type": "integer",
                        "example": 100,
                        "nullable": true
                    },
                    "view_count": {
                        "type": "integer",
                        "example": 250
                    },
                    "status": {
                        "type": "string",
                        "example": "active"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "category": {
                        "$ref": "#/components/schemas/Category"
                    },
                    "user": {
                        "$ref": "#/components/schemas/User"
                    }
                },
                "type": "object"
            },
            "User": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "John Doe",
                        "nullable": true
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "john@example.com"
                    },
                    "email_verified_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "avatar_url": {
                        "type": "string",
                        "example": "avatars/abc123.jpg",
                        "nullable": true
                    },
                    "role": {
                        "type": "string",
                        "example": "user"
                    },
                    "suspended_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "provider": {
                        "type": "string",
                        "example": "google",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Enter your Sanctum token. Get it by logging in via POST /api/auth/login or POST /api/auth/register.",
                "scheme": "bearer",
                "bearerFormat": "Sanctum"
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}